Convert fluorescence intensity (FI) to number of Photoelectrons (nPE)

CAUTION: Very experimental function. Channels of selected flow cytometers have been analyzed with a QuantiFlash device. By generation of very defined pulses of light the fluorescence intensity (FI) measured by a PMT (detector in a flow cytometer) can be correlated to an absolute number of detected photoelectrons (nPE). Currently this only works for fluorescence channels, not scatter channels. Also, strictly speaking, it is only valid for Height-channels (-H at the end)- For Area (-A) and Width (-W) it may not be correct. Fluorescence signals acquired at different voltages and even by different flow cytometers become comparable on the nPE-scale. A good idea would be to test the conversion by analyzing the same sample with different settings and/or at different machines.

nPE_to_fcs(
  file_path,
  h_channels_only = T,
  compensate = F,
  compMat,
  logicle_trans = F,
  kfactor_df,
  output_folder = NULL,
  new_file_suffix = "nPE"
)

Arguments

file_path
character, path to a fcs file

h_channels_only
logical, have only the height channels converted to nPE; height channels are actually the only ones for which the conversion is correct.

compensate
logical, should compensation be applied before PC calculation

compMat
matrix, optional; a compensation matrix to use for compensation. If not provided the SPILL argument of the fcs file will be used. If you have generated a compensation matrix in FlowJo see ?fcexpr::wsx_compMats_to_fcs in order to have it copied to fcs files.

logicle_trans
logical, should the logical transformation (Parks, 2006, https://pubmed.ncbi.nlm.nih.gov/16604519/) be applied before nPE calculation.

kfactor_df
data.frame, table with k-factors for every channel and voltage per machine, defaults to system.file("extdata", "k_factors.rds", package = "fcexpr")

output_folder
character, optional, path to a folder where to save the newly generated fcs file. Default is dirname(file_path).

new_file_suffix
character, the suffix to add to the the newly generated fcs file. Default is _nPE.

Value

appended flowFrame which is also saved as fcs file