A modified version of flowCores read.FCS

Please see ?flowCore::read.FCS for documentation. The additional feature of this modified version is the attempt to read compensation-matrices (SPILL keyword) which have been written with a comma as decimal separator instead of a dot. This error may occurs, when the language computer of BD Flow Cytometer is set German. In this case the DIVA software writes an FCS file which does not obey the international standard of using a dot as decimal separator. FlowJo by the way ignores this problem and just read the FCS file ignoring the SPILL keyword. In this case information is lost but the file can be used for further analyses. FlowCores read.FCS function though does not read these files.

read.FCS_comma(
  filename,
  transformation = "linearize",
  which.lines = NULL,
  alter.names = FALSE,
  column.pattern = NULL,
  invert.pattern = FALSE,
  decades = 0,
  ncdf = FALSE,
  min.limit = NULL,
  truncate_max_range = TRUE,
  dataset = NULL,
  emptyValue = TRUE,
  channel_alias = NULL,
  ...
)

Arguments

filename
see ?read.FCS

transformation
see ?read.FCS

which.lines
see ?read.FCS

alter.names
see ?read.FCS

column.pattern
see ?read.FCS

invert.pattern
see ?read.FCS

decades
see ?read.FCS

ncdf
see ?read.FCS

min.limit
see ?read.FCS

truncate_max_range
see ?read.FCS

dataset
see ?read.FCS

emptyValue
see ?read.FCS

channel_alias
see ?read.FCS

...
see ?read.FCS

Value

a flowFrame with corrected SPILL keyword

Details

In essence the function txt2spillmatrix tries to to guess which comma is a decimal separator and which one is a separator for different entries in the compensation matrix. One assumption is that the diagonal of the matrix has 1s only.

Examples

if (FALSE) {
# read the FCS file while the spill keyword is fixed (wrong comma replaced by dots)
ff <- read.FCS_spill.comma.sensitive(filename = "123.fcs")
# save the correct version as new file or overwrite the old one
# NOTE: the meta data reveal that the file was read and saved with flowCore
flowCore::write.FCS(x = ff, filename = "123_spill_corr.fcs")
}