Split a flowjo export of multiple plots into separate graphics for rearrangement

Split a flowjo export of multiple plots into separate graphics for rearrangement

split_flowjo_export_image(
  img,
  nPlots_x = 4,
  nPlots_y = 3,
  folder,
  pptx_name,
  pptx_image_size = 1,
  pptx_border_space = 0.2
)

Arguments

img
character path to an image, preferable a png exported from flowjos layout editor, has to be equally spaced in x and y direction

nPlots_x
integer, how many plots in x direction

nPlots_y
integer, how many plots in y direction

folder
folder name of where to save split images, this folder is written to the directory of img

pptx_name
character, name a pptx to write the split images to; if missing no pptx is generated

pptx_image_size
numeric, size of the images in pptx

pptx_border_space
numeric, space between images in pptx

Value

writes a folder with split images and optionally a pptx

Examples

if (FALSE) {
# copy an example file to your disk
file.copy(from = system.file("extdata", "img.png", package = "fcexpr"), to = "your_path")
fcexpr::split_flowjo_export_image(img = "your_path/img.png",
nPlots_x = 6, nPlots_y = 6, pptx = "my_img_split.pptx")
}