Retrieve MFIs from populations gated in one or more flowjo workspaces

If inverse.transform is set to TRUE values as shown by flowjo will be used for calculation. This may omit to use 'geo_mean' for mean.fun as negative values are not allowed. Inverse.transform = F leads to data transformation by the logicle transform.

ws_get_MFIs(
  ws,
  gr,
  population = NULL,
  FCS.file.folder = file.path(getwd(), "FCS_files"),
  mean.fun = "median",
  variance.fun = "sd",
  inverse.transform = T
)

Arguments

ws
character vector of paths to flowjo workspaces

gr
character vector of flowjo groups to import or a list of those vectors, one for each ws

population
character vector of which populations to calculate values for; if omitted calculation is done for all population which increases computational time

FCS.file.folder
path to root folder which contains FCS files; if missing file.path(getwd(), 'FCS_files') is assumed

mean.fun
character name of the function name to use to calculate the average FI

variance.fun
character name of the function name to use to calculate the variacnce of FI

inverse.transform
logical indicating if fluorescence values are to be inversely transformed

Value

data.frame

Examples

if (FALSE) {
# When the script is saved to R_scripts in the experiment folder,
# get the absolute path to the folder
wd <- dirname(dirname(rstudioapi::getActiveDocumentContext()$path))
# find workspaces
ws <- list.files(wd, pattern = '\\.wsp$', recursive = T, full.names = T)
# get groups
gr <- lapply(ws, function(x) {
unique(as.character(CytoML::fj_ws_get_sample_groups(CytoML::open_flowjo_xml(x))$groupName))
})[[1]]
# read the compensated values from flowjo and calculate MFIs
ws_get_MFIs(ws = ws, gr = gr)
}