Convert fluorescence intensity (FI) to number of detected photo electrons (nPE)

CAUTION: Very experimental function. Channels of selected flow cytometers have been analyzed with a QuantiFlash. By generation of very defined light pulses the fluorescence intensity (FI) measured by a PMT can be correlated to an absolute number of detected photo electrons (nPE). Currently this only works for fluorescence channels, not scatter channels. Also, strictly speaking, it is only valid for Height-channels (-H at the end)- For Area (-A) and Width (-W) it may not be correct. Fluorescence signals acquired at different voltages and even by different flow cytometers become comparable on the nPE-scale. A good idea would be to test the conversion by analyzing the same sample with different settings and/or at different machines.

nPE_to_fcs(
  file_path,
  h_channels_only = T,
  kfactor_df = NULL,
  output_folder = NULL,
  new_file_suffix = "nPE"
)

Arguments

file_path

character, path to a fcs file

h_channels_only

logical, have only the height channels converted to nPE; height channels are actually the only ones for which the conversion is correct.

kfactor_df

data.frame, table with k-factors for every channel and voltage per machine, defaults to system.file("extdata", "k_factors.rds", package = "fcexpr")

output_folder

character, optional, path to a folder where to save the newly generated fcs file. Default is dirname(file_path).

new_file_suffix

character, the suffix to add to the the newly generated fcs file. Default is _nPE.

Value

appended flowFrame which is also saved as fcs file to disk

Examples

if (FALSE) {
ff <- nPE_to_fcs(file_path = "myfolder/my_file.fcs")
}